refs #24 don't zap Em::Ssh::Shell callbacks
The ``@shell`` instance variable for the Em::Ssh::Shell instance is a
fresh ``Net::SSH::Connection::Channel`` instance that is extended with
``EM::Ssh::Connection::Channel::Interactive``, which in turn provides
callbacks to that instance. The change for #24 was setting the callbacks
for the ``Em::Ssh::Shell`` instance rather than the
``Net::SSH::Connection::Channel`` instance.